+++ module/function/auth/math.js
@@ -1,14 +1,14 @@
 module.exports = async(interaction)=>{
   const { ActionRowBuilder, ModalBuilder, TextInputBuilder } = require("discord.js");
   if(!interaction.isButton()) return;
-  if(interaction.customId.startsWith("panel_")){
+  if(interaction.customId.startsWith("math_")){
     const role = interaction.customId.split("_");
     const count_1 = Math.floor(Math.random()*15) + 1;
     const count_2 = Math.floor(Math.random()*15) + 1;
     const total = count_1 + count_2
 
     const check = new ModalBuilder()
-      .setCustomId(`panelrole_${role[1]}_${total}`)
+      .setCustomId(`mathrole_${role[1]}_${total}`)
       .setTitle("認証");
 
     const code = new TextInputBuilder()
